+++ INCLUDE/OPS/elemental_op.h
@@ -8,7 +8,7 @@
 struct ElementalOp
 {
 	public:
-		int coef;
+		double coef;
 		std::vector<Meson> mesons;
 
 		ElementalOp(int c, std::vector<Meson> n_m):mesons(n_m),coef(c){};
